# glotcast-api/openapi.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/lists/{slug}": {
      "get": {
        "operationId": "ListsController_page_v1",
        "parameters": [
          {
            "name": "slug",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListPageDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/v1/podcasts": {
      "get": {
        "operationId": "PodcastsController_list_v1",
        "parameters": [
          {
            "name": "category",
            "required": false,
            "in": "query",
            "description": "a category slug",
            "schema": {
              "maxLength": 80,
              "type": "string"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PodcastPageDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "catalog"
        ]
      }
    },
    "/v1/podcasts/{id}": {
      "get": {
        "operationId": "PodcastsController_detail_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID or legacy Strapi documentId",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PodcastDetailDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "catalog"
        ]
      }
    },
    "/v1/podcasts/{id}/episodes": {
      "get": {
        "operationId": "PodcastsController_episodes_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID or legacy Strapi documentId",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "sort",
            "required": false,
            "in": "query",
            "description": "by episode number",
            "schema": {
              "default": "asc",
              "type": "string",
              "enum": [
                "asc",
                "desc"
              ]
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpisodePageDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "catalog"
        ]
      }
    },
    "/v1/episodes": {
      "get": {
        "operationId": "EpisodesController_feed_v1",
        "parameters": [
          {
            "name": "feed",
            "required": false,
            "in": "query",
            "schema": {
              "default": "latest",
              "type": "string",
              "enum": [
                "latest",
                "following",
                "trending"
              ]
            }
          },
          {
            "name": "level",
            "required": false,
            "in": "query",
            "description": "only episodes that have this level",
            "schema": {
              "$ref": "#/components/schemas/Level"
            }
          },
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 9007199254740991,
              "default": 1,
              "type": "integer"
            }
          },
          {
            "name": "pageSize",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 50,
              "default": 20,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpisodePageDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "catalog"
        ]
      }
    },
    "/v1/episodes/{id}": {
      "get": {
        "operationId": "EpisodesController_detail_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID or legacy Strapi documentId",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EpisodeDetailDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "catalog"
        ]
      }
    },
    "/v1/episodes/{id}/transcript": {
      "get": {
        "operationId": "EpisodesController_transcript_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "UUID or legacy Strapi documentId",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "level",
            "required": true,
            "in": "query",
            "schema": {
              "$ref": "#/components/schemas/Level"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "catalog"
        ]
      }
    },
...
    "/v1/admin/podcasts/{id}": {
      "get": {
        "operationId": "AdminPodcastsController_get_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPodcastDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      },
      "patch": {
        "operationId": "AdminPodcastsController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdatePodcastDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminPodcastDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      },
      "delete": {
        "operationId": "AdminPodcastsController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
...
    "/v1/admin/episodes/{id}": {
      "get": {
        "operationId": "AdminEpisodesController_get_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminEpisodeDetailDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      },
      "patch": {
        "operationId": "AdminEpisodesController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateEpisodeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminEpisodeDetailDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      },
      "delete": {
        "operationId": "AdminEpisodesController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
...
    "/v1/admin/lists/{id}": {
      "get": {
        "operationId": "AdminListsController_get_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminListDetailDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      },
      "patch": {
        "operationId": "AdminListsController_update_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateListDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AdminListDetailDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      },
      "delete": {
        "operationId": "AdminListsController_remove_v1",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
...
    "/v1/admin/transcribe": {
      "post": {
        "operationId": "AdminTranscribeController_submit_v1",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TranscribeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscribeJobDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
    "/v1/admin/transcribe/{jobId}": {
      "get": {
        "operationId": "AdminTranscribeController_get_v1",
        "parameters": [
          {
            "name": "jobId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TranscriptionDto_Output"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "admin"
        ]
      }
    },
...
  },
  "info": {
    "title": "GlotCast API",
    "description": "Podcasts in three English levels: catalog, listening, vocabulary, translation, and the admin content pipeline. The contract is docs/contract-v1.md; `admin`-tagged operations are for the admin panel only.",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      }
    },
    "schemas": {
...
      "ListPageDto_Output": {
        "type": "object",
        "properties": {
          "list": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string",
                "format": "uuid",
                "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
              },
              "slug": {
                "type": "string"
              },
              "name": {
                "type": "string"
              },
              "description": {
                "anyOf": [
                  {
                    "type": "string",
                    "minLength": 0
                  },
                  {
                    "type": "null"
                  }
                ]
              }
            },
            "required": [
              "id",
              "slug",
              "name",
              "description"
            ],
            "additionalProperties": false
          },
          "episodes": {
            "type": "object",
            "properties": {
              "items": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/EpisodeSummary_Output"
                }
              },
              "page": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "pageSize": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "total": {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              "hasMore": {
                "type": "boolean"
              }
            },
            "required": [
              "items",
              "page",
              "pageSize",
              "total",
              "hasMore"
            ],
            "additionalProperties": false
          }
        },
        "required": [
          "list",
          "episodes"
        ],
        "additionalProperties": false
      },
...
      "Level": {
        "type": "string",
        "enum": [
          "bg",
          "in",
          "ad"
        ],
        "description": "bg = beginner, in = intermediate, ad = advanced"
      },
...
      "PodcastPageDto_Output": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PodcastSummary_Output"
            }
          },
          "page": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "pageSize": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "total": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "hasMore": {
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total",
          "hasMore"
        ],
        "additionalProperties": false
      },
...
      "PodcastDetailDto_Output": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "slug": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "coverUrl": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "episodeCount": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991,
            "description": "published episodes"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "categories": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CategoryRef_Output"
            }
          },
          "isFollowing": {
            "type": "boolean",
            "description": "false without a token"
          },
          "levels": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Level_Output"
            },
            "description": "levels available across its episodes"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "coverUrl",
          "episodeCount",
          "description",
          "categories",
          "isFollowing",
          "levels"
        ],
        "additionalProperties": false
      },
      "EpisodePageDto_Output": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EpisodeSummary_Output"
            }
          },
          "page": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "pageSize": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "total": {
            "type": "integer",
            "minimum": -9007199254740991,
            "maximum": 9007199254740991
          },
          "hasMore": {
            "type": "boolean"
          }
        },
        "required": [
          "items",
          "page",
          "pageSize",
          "total",
          "hasMore"
        ],
        "additionalProperties": false
      },
      "EpisodeDetailDto_Output": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "podcast": {
            "$ref": "#/components/schemas/PodcastSummary_Output"
          },
          "number": {
            "anyOf": [
              {
                "type": "integer",
                "minimum": -9007199254740991,
                "maximum": 9007199254740991
              },
              {
                "type": "null"
              }
            ]
          },
          "title": {
            "type": "string"
          },
          "coverUrl": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "bannerUrl": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "isPro": {
            "type": "boolean"
          },
          "publishedAt": {
            "type": "string"
          },
          "levels": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "level": {
                  "$ref": "#/components/schemas/Level_Output"
                },
                "durationSec": {
                  "type": "number"
                },
                "description": {
                  "anyOf": [
                    {
                      "type": "string",
                      "minLength": 0
                    },
                    {
                      "type": "null"
                    }
                  ]
                },
                "audioUrl": {
                  "type": "string"
                }
              },
              "required": [
                "level",
                "durationSec",
                "description",
                "audioUrl"
              ],
              "additionalProperties": false
            }
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 0
              },
              {
                "type": "null"
              }
            ]
          },
          "isFavorite": {
            "type": "boolean"
          },
          "progress": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LevelProgress_Output"
            },
            "description": "the caller's per-level progress, [] for none"
          }
        },
        "required": [
          "id",
          "podcast",
          "number",
          "title",
          "coverUrl",
          "bannerUrl",
          "isPro",
          "publishedAt",
          "levels",
          "description",
          "isFavorite",
          "progress"
        ],
        "additionalProperties": false
      },
...
      "TranscriptDto_Output": {
        "type": "object",
        "properties": {
          "episodeId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          },
          "level": {
            "$ref": "#/components/schemas/Level_Output"
          },
          "chunks": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TranscriptChunk_Output"
            }
          }
        },
        "required": [
          "episodeId",
          "level",
          "chunks"
        ],
        "additionalProperties": false
      },
...
      "UpdateListDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "minLength": 1,
            "maxLength": 120
          },
          "slug": {
            "description": "default: from the name",
            "type": "string",
            "minLength": 1,
            "maxLength": 80,
            "pattern": "^[a-z0-9]+(?:-[a-z0-9]+)*$"
          },
          "description": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 2000
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
...
      "TranscribeDto": {
        "type": "object",
        "properties": {
          "audioUrl": {
            "type": "string",
            "maxLength": 2000,
            "format": "uri",
            "description": "a public audio URL (R2)"
          }
        },
        "required": [
          "audioUrl"
        ]
      },
      "TranscribeJobDto_Output": {
        "type": "object",
        "properties": {
          "jobId": {
            "type": "string",
            "format": "uuid",
            "pattern": "^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$"
          }
        },
        "required": [
          "jobId"
        ],
        "additionalProperties": false
      },
...
    }
  }
}